fix tests

+++ scripts/build.sh
@@ -1,14 +1,19 @@
 #!/bin/bash
 
-## This builds the whl and then copies it to all 4 test charms and updates the requirements file.
-
 set -e
 
-git_hash=$(git describe --always --dirty)
-
 LIB_PATH="./single_kernel_opensearch_dashboards"
-
+SPECIAL_CHARM="tests/integration/application_charm"
 CHARMS_PATH="./tests/charms"
+
+# Helper function to avoid code duplication
+pack_charm() {
+    if ${CI_CACHE:-false}; then
+        ccc pack -v
+    else
+        charmcraft pack -v
+    fi
+}
 
 if [ $# -ge 1 ]; then
     declare -a TEST_CHARMS=("$1")
@@ -17,6 +22,18 @@ else
 fi
 
 for directory in "${TEST_CHARMS[@]}"; do
+    # Check if the current directory matches the special charm path
+    # We use wildcard * to catch it even if the user passes "./tests/..."
+    if [[ "$directory" == *"$SPECIAL_CHARM" ]]; then
+        printf 'Building charm %s \n'"${directory}"
+        pushd "$directory"
+        pack_charm
+        popd
+
+        # Skip the rest of the loop for this iteration
+        continue
+    fi
+
     echo "clearing out libs for charm"
     directory_lib_path="${directory}/${LIB_PATH}"
     rm -rf "$directory_lib_path"
@@ -26,10 +43,10 @@ for directory in "${TEST_CHARMS[@]}"; do
     cp "pyproject.toml" "$directory_lib_path"
     cp "README.md" "$directory_lib_path"
 
-    echo "Building charm ${directory}\n"
+    printf 'Building charm %s \n'"${directory}"
 
 
-    pushd $directory
+    pushd "$directory"
 
     # Backup files
     cp pyproject.toml pyproject.toml.backup
@@ -47,11 +64,7 @@ for directory in "${TEST_CHARMS[@]}"; do
     python3 -c 'import pathlib; import shutil; import subprocess; git_hash=subprocess.run(["git", "describe", "--always", "--dirty"], capture_output=True, check=True, encoding="utf-8").stdout; file = pathlib.Path("charm_version"); shutil.copy(file, pathlib.Path("charm_version.backup")); version = file.read_text().strip(); file.write_text(f"{version}+{git_hash}")'
 
     # Pack the charm
-    if $CI_CACHE; then
-        ccc pack -v
-    else
-        charmcraft pack -v
-    fi
+    pack_charm
 
     # Cleanup
     echo "removing copied files from single kernel charm."
